==> BDCI2017LZSJ/feature/0_analysi_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns


def spliteData(df_proData):
    """
    对数据df_proData,分割成训练和测试数据集
    并给训练数据加上标签和测试数据分割开
    """
    df_label = pd.read_csv('../data/public/train.csv')

    df_train = df_proData[df_proData['EID'].isin(df_label['EID'])]
    df_test = df_proData[~df_proData['EID'].isin(df_label['EID'])]
    
    
    df_train = pd.merge(df_train,df_label[df_label['EID'].isin(df_train['EID'])],how ='left',on=['EID'])
    
    return df_train,df_test


df_train = pd.read_csv('../data/public/train.csv')
df_branch = pd.read_csv('../data/public/3branch_1.csv')


df_train = df_branch[df_branch['EID'].isin(df_label['EID'])]
df_test = df_branch[~df_branch['EID'].isin(df_label['EID'])]


df_label.info()
df_label.sort_values(['EID'])

df_train = pd.merge(df_train,df_label[df_label['EID'].isin(df_train['EID'])],how ='left',on=['EID'])

df_train.info()

df_train.sort_values(['TARGET'],ascending=0)

df_label[df_label['EID']==46163]
