==> BDCI2017LZSJ/feature/1_feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt


# 得到所有的企业本身的特征
    0. 企业注册时间差，RGYEAR
    1. 2018-企业注册时间差，RGYEAR_DIFF
    2. 注册资本，ZCZB
    3. 企业的行业大类，HY
    4. 企业的类型，ETYPE
    5. 注册资本占整个行业大类的注册资本的平均值的比例，ZCZB_RATE
    6. 注册资本占这个企业类型的注册资本的平均值的比例，ETYPE_RATE
    
    7. 已经完成的企业的计算指标，['MPNUM', 'INUM', 'FINZB','FSTINUM', 'TZINUM'],用0填充
    -----------------------------------------------------------------------------------------
    8. 每个企业对应行业大类的平均资金，HY_ZCZB_AVG
    9. 每个企业对应行业大类的平均资金占所有企业的平均注册资金的比例, HY_ZCZB_AVG_RATE

    8. 每个企业对应行业大类的平均资金，ETYPE_ZCZB_AVG
    9. 每个企业对应行业大类的平均资金占所有企业的平均注册资金的比例, ETYPE_ZCZB_AVG_RATE
    
        
    10. 每个完成的企业的计算指标其对应行业大类的计算指标的平均值,['MPNUM_HY_AVG', 'INUM_HY_AVG', 'FINZB_HY_AVG','FSTINUM_HY_AVG', 'TZINUM_HY_AVG']
    
    
        ['MPNUM_HY_AVG', 'INUM_HY_AVG', 'FINZB_HY_AVG','FSTINUM_HY_AVG', 'TZINUM_HY_AVG'] / MPNUM_AVG ... 
    11. 每个完成的企业的计算指标其对应行业大类的计算指标的平均值占整个计算指标的平均值的比例,['MPNUM_HY_AVG_RATE', 'INUM_HY_AVG_RATE', 'FINZB_HY_AVG_RATE','FSTINUM_HY_AVG_RATE', 'TZINUM_HY_AVG_RATE']
        
        
        ['MPNUM', 'INUM', 'FINZB','FSTINUM', 'TZINUM'] / ['MPNUM_HY_AVG', 'INUM_HY_AVG', 'FINZB_HY_AVG','FSTINUM_HY_AVG', 'TZINUM_HY_AVG']
    12. 每个完成的企业的计算指标占其对应行业大类对应的计算指标的平均值的比例,['MPNUM_HY_RATE', 'INUM_HY_RATE', 'FINZB_HY_RATE','FSTINUM_HY_RATE', 'TZINUM_HY_RATE']
     

    


company = pd.read_csv('../python_workplace/BDCI2017LZSJ/data/public/1entbase.csv')


company['ETYPE'].value_counts()

company = company.fillna(0)
company.info()
company.head()

tmp = pd.DataFrame()
tmp['HY_ZCZB_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['ZCZB'].mean()
tmp['HY_ZCZB_AVG_RATE'] = tmp['HY_ZCZB_AVG'] / company['ZCZB'].mean()

tmp['MPNUM_HY_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['MPNUM'].mean()
tmp['INUM_HY_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['INUM'].mean()
tmp['FINZB_HY_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['FINZB'].mean()
tmp['FSTINUM_HY_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['FSTINUM'].mean()
tmp['TZINUM_HY_AVG'] = company.groupby(['HY'],as_index=True,axis=0)['TZINUM'].mean()


tmp['HY'] = tmp.index

tmp1 = pd.DataFrame()
tmp1['ETYPE_ZCZB_AVG'] = company.groupby(['ETYPE'],as_index=True,axis=0)['ZCZB'].mean()
tmp1['ETYPE_ZCZB_AVG_RATE'] = tmp1['ETYPE_ZCZB_AVG'] / company['ZCZB'].mean()



tmp1['ETYPE'] = tmp1.index



company = pd.merge(company,tmp,how='left',on=['HY'])
company = pd.merge(company,tmp1,how='left',on=['ETYPE'])


company.info()
company.head()

company['RGYEAR_DIFF'] = 2018 - company['RGYEAR']
company['ZCZB_RATE'] = company['ZCZB'] / company['HY_ZCZB_AVG']
company['ETYPE_RATE'] = company['ZCZB'] / company['ETYPE_ZCZB_AVG']

company['MPNUM_HY_AVG_RATE'] = company['MPNUM_HY_AVG'] / company['MPNUM'].mean()
company['INUM_HY_AVG_RATE'] = company['INUM_HY_AVG'] / company['INUM'].mean()
company['FINZB_HY_AVG_RATE'] = company['FINZB_HY_AVG'] / company['FINZB'].mean()
company['FSTINUM_HY_AVG_RATE'] = company['FSTINUM_HY_AVG'] / company['FSTINUM'].mean()
company['TZINUM_HY_AVG_RATE'] = company['TZINUM_HY_AVG'] / company['TZINUM'].mean()


company['MPNUM_HY_RATE'] = company['MPNUM'] / company['MPNUM_HY_AVG']
company['INUM_HY_RATE'] = company['INUM'] / company['INUM_HY_AVG']
company['FINZB_HY_RATE'] = company['FINZB'] / company['FINZB_HY_AVG']
company['FSTINUM_HY_RATE'] = company['FSTINUM'] / company['FSTINUM_HY_AVG']
company['TZINUM_HY_RATE'] = company['TZINUM'] / company['TZINUM_HY_AVG']

company = company.fillna(0)

def setTime(x):
    return str(x)+'-01'

company['RGYEAR'] = company['RGYEAR'].apply(setTime)



company.info()
company.head()

HY的类别更多
FSTINUM、INUM、MPNUM、FINZB:用同一个企业类型（HY）的平均值填充空值（取整）
TZINUM：用这个HY中出现最多的值进行填充


company.to_csv('../data/alldata/df_data1.csv',index=False,index_label=False)
