==> BDCI2017LZSJ/feature/2_feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import re
import pro_data

# 企业变更数据的对应特征(2alter.csv)
    
    1. 变更的事件的总个数, ALTERNO_SUM
    2. 变更的事件的类型的种类数, ALTERNO_CNT
    3. 变更的事件的类型的种类数 / 所有事件的总总的类别数, ALTERNO_CNT_RATE
    4. 第一次变更时间, FIRST_CHANGE_TIME
    5. 第一次变更距离企业注册时间的差（单位：月）,不存在变更事件的用0填充, FIRST_CHANGE_TIME_DIFF
    6. 最后一次变更时间, END_CHANGE_TIME
    7. 最后一次变更时间距离企业注册时间的差值，END_CHANGE_TIME_DIFF
    8. 最后一次变更距离第一次事件变更时间的差（单位：月）,不存在变更事件的用0填充, END_FIRST_CHANGE_TIME_DIFF

    9. 变更的事件是否发生了前后的值得变化（1/0）, IS_CHANGE
    10. 第一次变更事件的类型，CHANGE_FIRST_ALTERNO
    11. 最后一次变更事件的类型，CHANGE_END_ALTERNO
    12. 最多的变更事件的类型，CHANGE_ALTERNO_MUCNID
    
    新加特征
    -------------------------------------------------------------------------------------------------------
    13. 平均每几个月变更一次事件，CHANGE_PRE_CNT = ALTERNO_SUM / END_CHANGE_TIME_DIFF
    14. 每个企业变更事件的个数占所有企业变更事件平均值的比例， CHANGE_CNT_RATE = ALTERNO_SUM / AVG 
    15. 企业对应大类HY的平均变更次数，CHANGE_HY_CNT_AVG
    16. 企业对应大类HY的平均变更次数占所有变更时间的个数的平均值的比例，CHANGE_HY_CNT_ALL_AVG




df_all = pd.read_csv("../data/alldata/df_data1.csv")
df_change = pd.read_csv("../data/public/2alter.csv")

df_all['ETYPE'].value_counts()

df_change.info()
df_change.head()

df_all.info()

df_change = df_change.sort_values(['EID','ALTDATE'])


df_change = df_change.fillna('0')
columns = ['EID', 'ALTERNO', 'ALTDATE', 'ALTBE', 'ALTAF']
df_xchange = pd.DataFrame(columns=['EID', 'ALTERNO', 'ALTDATE', 'ALTBE', 'ALTAF'])

EIDS = set(df_change['EID'])

len(EIDS)

k = 0
for EID in EIDS:
    if k%3000 == 0:
        print('第%d次处理--------->',k)
    k+=1
    tmp = df_change[df_change['EID'] == EID]
    row = [EID,tmp['ALTERNO'].values,tmp['ALTDATE'].values,tmp['ALTBE'].values,tmp['ALTAF'].values]
    
    df_xchange = df_xchange.append(pd.Series(row,columns),ignore_index=True)
    

df_xchange.info()
df_xchange.head()



set(df_change['ALTERNO'])

def getFAlterno(x):
    m = x[0]
    if m == 'A_015':
        m = m[3:]
    return m

def getEAlterno(x):
    m = x[-1]
    if m == 'A_015':
        m = m[3:]
    return m

def getMAlterno(x):
    x = list(x)
    m = x[0]
    k = 0
    for i in x:
        if k<x.count(i):
            k=x.count(i)
            m = i
    if m == 'A_015':
        m = m[3:]
    return m


df_xchange['CHANGE_FIRST_ALTERNO'] = df_xchange['ALTERNO'].apply(getFAlterno)
df_xchange['CHANGE_END_ALTERNO'] = df_xchange['ALTERNO'].apply(getEAlterno)
df_xchange['CHANGE_ALTERNO_MUCNID'] = df_xchange['ALTERNO'].apply(getMAlterno)


df_xchange.head()

k = len(set(df_change['ALTERNO']))

def totalcnt(x):
    return len(x)

df_xchange['ALTERNO_SUM'] = df_xchange['ALTERNO'].apply(totalcnt)

def cnt(x):
    return len(set(x))

df_xchange['ALTERNO_CNT'] = df_xchange['ALTERNO'].apply(cnt)

df_xchange['ALTERNO_CNT_RATE'] = df_xchange['ALTERNO_CNT'] / k

def getFirst(x):
    x.sort()
    return x[0]

df_xchange['FIRST_CHANGE_TIME'] = df_xchange['ALTDATE'].apply(getFirst)

def getEnd(x):
    x.sort()
    return x[-1]

df_xchange['END_CHANGE_TIME'] = df_xchange['ALTDATE'].apply(getEnd)

def equal(x):
    ok = 0
    for d in x:
        if d[:d.find(":")] != d[d.find(":")+1:]:
            ok = 1
    return ok
df_xchange['IS_CHANGE'] = (df_xchange['ALTBE'] +':'+ df_xchange['ALTAF']).apply(equal)

df_xchange.to_csv('../data/public/2alter_1.csv',index=False,index_label=False)

# df_xchange = pd.read_csv('../data/public/2alter_1.csv')

df_all = pd.merge(df_all,df_xchange[['EID', 'ALTERNO_SUM','ALTERNO_CNT', 'ALTERNO_CNT_RATE', 'FIRST_CHANGE_TIME',
                            'END_CHANGE_TIME', 'IS_CHANGE','CHANGE_FIRST_ALTERNO','CHANGE_END_ALTERNO','CHANGE_ALTERNO_MUCNID']],how='left',on=['EID'])

df_all.info()
df_all.head()

> 空值填充，根据HY的类别的平均值或者众数进行填充
   + ['ALTERNO_SUM','ALTERNO_CNT', 'ALTERNO_CNT_RATE',]使用同一个大类别的均值进行填充
   + ['FIRST_CHANGE_TIME','END_CHANGE_TIME', 'IS_CHANGE' ]使用同一个大类别的众数进行填充



HYLIST = set(df_all['HY'])

print(HYLIST)

meanlist = ['ALTERNO_SUM','ALTERNO_CNT', 'ALTERNO_CNT_RATE']
modelist = ['FIRST_CHANGE_TIME','END_CHANGE_TIME', 'IS_CHANGE','CHANGE_FIRST_ALTERNO','CHANGE_END_ALTERNO','CHANGE_ALTERNO_MUCNID' ]



int(df_all[df_all['HY']==75]['ALTERNO_SUM'].mean())

df_all[df_all['HY']==7]['CHANGE_FIRST_ALTERNO'].value_counts().index[0]



for HY in HYLIST:
    # print(df_train['HY'].value_counts())
    for d in meanlist:
        df_all.loc[df_all[df_all[d].isnull()][df_all['HY']==HY].index,d] = df_all[df_all['HY']==HY][d].mean()
    
    for c in modelist:
        if(len(df_all[df_all['HY']==HY][c].value_counts().index)==0): continue
        df_all.loc[df_all[df_all[c].isnull()][df_all['HY']==HY].index,c] = df_all[df_all['HY']==HY][c].value_counts().index[0]


# df_all[df_all['FIRST_CHANGE_TIME'].isnull()]['FIRST_CHANGE_TIME'] = df_all[df_all['FIRST_CHANGE_TIME'].isnull()]['RGYEAR'] 

# df_all.loc[df_all['FIRST_CHANGE_TIME'].isnull(),'FIRST_CHANGE_TIME'] = df_all.loc[df_all['FIRST_CHANGE_TIME'].isnull(),'RGYEAR']
# df_all.loc[df_all['END_CHANGE_TIME'].isnull(),'END_CHANGE_TIME'] = df_all.loc[df_all['END_CHANGE_TIME'].isnull(),'RGYEAR']

def timeDiff(x):
    a = x[:x.find(':')]
    b = x[x.find(':')+1:]
    y = int(a[:a.find('-')]) - int(b[:b.find('-')])
    m = int(a[a.find('-')+1:]) - int(b[b.find('-')+1:])
    return y * 12 + m
    
df_all['FIRST_CHANGE_TIME_DIFF'] = (df_all['FIRST_CHANGE_TIME'] + ':' + df_all['RGYEAR']).apply(timeDiff)
df_all['END_CHANGE_TIME_DIFF'] = (df_all['END_CHANGE_TIME'] + ':' + df_all['RGYEAR']).apply(timeDiff)

df_all['END_FIRST_CHANGE_TIME_DIFF'] = (df_all['END_CHANGE_TIME'] + ':' + df_all['FIRST_CHANGE_TIME']).apply(timeDiff)


df_all.info()

13. 平均每几个月变更一次事件，CHANGE_PRE_CNT = ALTERNO_SUM / END_CHANGE_TIME_DIFF
    14. 每个企业变更事件的个数占所有企业变更事件平均值的比例， CHANGE_CNT_RATE = ALTERNO_SUM / AVG 
    15. 企业对应大类HY的平均变更次数，CHANGE_HY_CNT_AVG
    16. 企业对应大类HY的平均变更次数占所有变更时间的个数的平均值的比例，CHANGE_HY_CNT_ALL_AVG


df_all['CHANGE_PRE_CNT'] = df_all['ALTERNO_SUM'] / df_all['END_CHANGE_TIME_DIFF']
df_all['CHANGE_CNT_RATE'] = df_all['ALTERNO_SUM'] / df_all['ALTERNO_SUM'].mean()


tmp = pd.DataFrame()

tmp['CHANGE_HY_CNT_AVG'] = df_all.groupby(['HY'],as_index=True,axis=0)['ALTERNO_SUM'].mean()
tmp['CHANGE_HY_CNT_ALL_AVG'] = tmp['CHANGE_HY_CNT_AVG'] / df_all['ALTERNO_SUM'].mean()
tmp['HY'] = tmp.index





df_all1.info()

df_all = pd.merge(df_all,tmp,how='left',on=['HY'])


df_all.info()
df_all.head()

df_all.to_csv('../data/alldata/df_data12.csv',index=False,index_label=False)

df_all.info()

# 得到在df_xchange的所有数据，
# df_xall = df_all[df_all['EID'].isin(df_xchange['EID'])]


# 分成训练和测试文件写入文件
# df_train,df_test = pro_data.spliteData(df_xall)
# df_train.to_csv('../data/newData/df_train01.csv',index=False,index_label=False)
# df_test.to_csv('../data/newData/df_test01.csv',index=False,index_label=False)

# df_all.to_csv('../data/alldata/df_data01.csv',index=False,index_label=False)
