==> BDCI2017LZSJ/feature/6_feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import pro_data

+ 第一个修改特征
# 项目数据6project.csv提取的特征
    1. 项目的数量个数， PROJECT_CNT,
    2. 省内项目的个数， PROJECT_INHOME_CNT
    3. 省外项目的个数， PROJECT_OUTHOME_CNT
    4. 省内项目的比例， PROJECT_INHOME_RATE
    5. 省外项目的比例， PROJECT_OUTHOME_RATE
    
    6. 项目的类型数目，PROJECT_TYPECODE_CNT

    7. 第一个项目的项目ID，PROJECT_FIRST_TYPECODE
    8. 最后一个项目的项目ID，PROJECT_END_TYPECODE
    
    9. 第一个项目中标的时间，PROJECT_FIRST_TIME
    10. 最后一个项目中标的时间，PROJECT_END_TIME
    
    11. 最后一个项目中标时间和第一个中标时间的差值，PROJECT_END_FIRST_DIFF
    12. 第一个项目中标时间距离企业注册时间的差值，PROJECT_FIRST_RGYEAR_DIFF
    13. 最后一个项目中标时间距离企业注册时间的差值，PROJECT_END_RGYEAR_DIFF


    14. 第一个项目中标时间距离企业第一个变更时间的差值，PROJECT_FIRST_CHANGE_FIRST_DIFF
    15. 最后一个项目中标时间距离企第一个变更时间的差值，PROJECT_END_CHANGE_FIRST_DIFF
    16. 第一个项目中标时间距离企业最后一个变更时间的差值，PROJECT_FIRST_CHANGE_END_DIFF
    17. 最后一个项目中标时间距离企最后一个变更时间的差值，PROJECT_END_CHANGE_END_DIFF
    
    ---------------------------------------------------------------------------------
    18. 平均没几个月中标一个项目其余用0表示，PROJECT_PRE_MONTH_CNT = PROJECT_END_RGYEAR_DIFF / PROJECT_CNT

    19. 企业权利的个数占所有权利个数平均值的比例。PROJECT_CNT_ALL_RATE
    
    20. 企业对应的大类HY的平均权利的个数，PROJECT_HY_CNT_AVG
    21. 企业对应大类HY的平均权利个数占所有权利平均个数的比例，PROJECT_HY_CNT_ALL_RATE
    22. 企业权利的个数占其对应的大类HY的平均值的比例，PROJECT_CNT_HY_RATE

    
    

df_all = pd.read_csv("../data/alldata/df_data12345.csv")
df_project = pd.read_csv("../data/public/6project.csv")

df_all.info()
df_all.head()

df_project.info()
df_project.head()

len(df_project['TYPECODE'].value_counts())

EIDS = set(df_project['EID'])

print(len(EIDS))

columns = df_project.columns
df_xproject = pd.DataFrame(columns=columns)

print(columns)

k = 0
for EID in EIDS:
    if k%3000 == 0:
        print('第%d次处理--------->',k)
    k+=1
    tmp = df_project[df_project['EID'] == EID]
    row = [EID,tmp['TYPECODE'].values,tmp['DJDATE'].values,tmp['IFHOME'].values]
    
    df_xproject = df_xproject.append(pd.Series(row,columns),ignore_index=True)
    

df_xproject.info()
df_xproject.head()

df_xproject['PROJECT_CNT'] = df_xproject['TYPECODE'].apply(lambda x: len(x))
df_xproject['PROJECT_INHOME_CNT'] = df_xproject['IFHOME'].apply(lambda x: list(x).count(1))
df_xproject['PROJECT_OUTHOME_CNT'] = df_xproject['IFHOME'].apply(lambda x: list(x).count(0))


df_xproject['PROJECT_INHOME_RATE'] = df_xproject['PROJECT_INHOME_CNT'] / df_xproject['PROJECT_CNT']
df_xproject['PROJECT_OUTHOME_RATE'] = df_xproject['PROJECT_OUTHOME_CNT'] / df_xproject['PROJECT_CNT']



df_xproject['PROJECT_TYPECODE_CNT'] = df_xproject['TYPECODE'].apply(lambda x: len(set(x)) )

df_xproject['PROJECT_FIRST_TYPECODE'] = df_xproject['TYPECODE'].apply(lambda x: x[0])
df_xproject['PROJECT_END_TYPECODE'] = df_xproject['TYPECODE'].apply(lambda x: x[-1])

df_xproject['PROJECT_FIRST_TIME'] = df_xproject['DJDATE'].apply(lambda x: x[0])
df_xproject['PROJECT_END_TIME'] = df_xproject['DJDATE'].apply(lambda x: x[-1])


df_xproject.to_csv('../data/public/6project_1.csv',index=False,index_label=False)
df_xproject.columns



df_all = pd.merge(df_all,df_xproject[['EID', 'PROJECT_CNT','PROJECT_INHOME_CNT', 'PROJECT_OUTHOME_CNT', 'PROJECT_INHOME_RATE',
                                    'PROJECT_OUTHOME_RATE', 'PROJECT_TYPECODE_CNT', 'PROJECT_FIRST_TYPECODE', 'PROJECT_END_TYPECODE',
                                    'PROJECT_FIRST_TIME','PROJECT_END_TIME']],how='left',on=['EID'])


df_all.info()
df_all.head()

# 空值填充，根据HY的类别的平均值或者众数进行填充
# ['RIGHT_CNT','RIGHT_TYPE_CNT', 'RIGHT_TYPE_RATE']使用同一个大类别的均值进行填充
# ['RIGHT_FIRST_TYPECODE','RIGHT_END_TYPECODE', 'RIGHT_TYPECODE_MUCHID', 'RIGHT_FIRST_ASK_TIME',
#     'RIGHT_FIRST_FB_TIME', 'RIGHT_END_ASK_TIME', 'RIGHT_END_FB_TIME']使用同一个大类别的众数进行填充

HYLIST = set(df_all['HY'])
print(HYLIST)

meanlist = ['PROJECT_FIRST_TYPECODE', 'PROJECT_END_TYPECODE','PROJECT_CNT','PROJECT_INHOME_CNT', 'PROJECT_OUTHOME_CNT','PROJECT_OUTHOME_CNT','PROJECT_OUTHOME_RATE','PROJECT_TYPECODE_CNT']
modelist = ['PROJECT_FIRST_TIME', 'PROJECT_END_TIME']


for HY in HYLIST:
    # print(df_train['HY'].value_counts())
    for d in meanlist:
        df_all.loc[df_all[df_all[d].isnull()][df_all['HY']==HY].index,d] = df_all[df_all['HY']==HY][d].mean()
    
    for c in modelist:
        if(len(df_all[df_all['HY']==HY][c].value_counts().index)==0): continue
        df_all.loc[df_all[df_all[c].isnull()][df_all['HY']==HY].index,c] = df_all[df_all['HY']==HY][c].value_counts().index[0]


for d in meanlist:
    df_all.loc[df_all[d].isnull(),d] = df_all[d].mean()

for c in modelist:
    df_all.loc[df_all[c].isnull(),c] = df_all[c].value_counts().index[0]

df_all = df_all.fillna(0)

# 所有的关于时间的空值用2018-01填充，假设这些企业被执行的数据发生在下一年，其他用0填充便是没有被执行

for d in meanlist:
    df_all.loc[df_all[d].isnull(),d] = 0

for c in modelist:
    df_all.loc[df_all[c].isnull(),c] = '2018-01'


def timeDiff(x):
    a = x[:x.find(':')]
    b = x[x.find(':')+1:]
    y = int(a[:a.find('-')]) - int(b[:b.find('-')])
    m = int(a[a.find('-')+1:]) - int(b[b.find('-')+1:])
    return y * 12 + m


df_all['PROJECT_END_FIRST_DIFF'] = (df_all['PROJECT_END_TIME'] + ':' + df_all['PROJECT_FIRST_TIME']).apply(timeDiff)

df_all['PROJECT_FIRST_RGYEAR_DIFF'] = (df_all['PROJECT_FIRST_TIME'] + ':' + df_all['RGYEAR']).apply(timeDiff)
df_all['PROJECT_END_RGYEAR_DIFF'] = (df_all['PROJECT_END_TIME'] + ':' + df_all['RGYEAR']).apply(timeDiff)

df_all['PROJECT_FIRST_CHANGE_FIRST_DIFF'] = (df_all['PROJECT_FIRST_TIME'] + ':' + df_all['FIRST_CHANGE_TIME']).apply(timeDiff)
df_all['PROJECT_END_CHANGE_FIRST_DIFF'] = (df_all['PROJECT_END_TIME'] + ':' + df_all['FIRST_CHANGE_TIME']).apply(timeDiff)

df_all['PROJECT_FIRST_CHANGE_END_DIFF'] = (df_all['PROJECT_FIRST_TIME'] + ':' + df_all['END_CHANGE_TIME']).apply(timeDiff)
df_all['PROJECT_END_CHANGE_END_DIFF'] = (df_all['PROJECT_END_TIME'] + ':' + df_all['END_CHANGE_TIME']).apply(timeDiff)


    ---------------------------------------------------------------------------------
    18. 平均没几个月中标一个项目其余用0表示，PROJECT_PRE_MONTH_CNT = PROJECT_END_RGYEAR_DIFF / PROJECT_CNT

    19. 企业权利的个数占所有权利个数平均值的比例。PROJECT_CNT_ALL_RATE
    
    20. 企业对应的大类HY的平均权利的个数，PROJECT_HY_CNT_AVG
    21. 企业对应大类HY的平均权利个数占所有权利平均个数的比例，PROJECT_HY_CNT_ALL_RATE
    22. 企业权利的个数占其对应的大类HY的平均值的比例，PROJECT_CNT_HY_RATE


df_all['PROJECT_PRE_MONTH_CNT'] = df_all['PROJECT_END_RGYEAR_DIFF'] / df_all['PROJECT_CNT']
df_all['PROJECT_CNT_ALL_RATE'] = df_all['PROJECT_CNT'] / df_all['PROJECT_CNT'].mean()


tmp = pd.DataFrame()

tmp['PROJECT_HY_CNT_AVG'] = df_all.groupby(['HY'],as_index=True,axis=0)['PROJECT_CNT'].mean()
tmp['PROJECT_HY_CNT_ALL_RATE'] = tmp['PROJECT_HY_CNT_AVG'] / df_all['PROJECT_CNT'].mean()

tmp['HY'] = tmp.index

df_all = pd.merge(df_all,tmp,how='left',on=['HY'])



df_all['PROJECT_CNT_HY_RATE'] = df_all['PROJECT_CNT'] / df_all['PROJECT_HY_CNT_AVG']


df_all.info()
df_all.head()


# 得到在df_xchange的所有数据，
df_all = df_all.fillna(0)
df_all.to_csv('../data/alldata/df_data123456.csv',index=False,index_label=False)
